Move some functions to vtk_utils
