+++ vibra/interface/viewer_3d/actors/multimaterial.py
@@ -1,5 +1,4 @@
 from collections import defaultdict
-from pathlib import Path
 from typing import Sequence
 
 import numpy as np
@@ -10,7 +9,6 @@ from vtkmodules.util.numpy_support import (
     vtk_to_numpy,
 )
 from vtkmodules.vtkCommonCore import (
-    vtkIdList,
     vtkPoints,
 )
 from vtkmodules.vtkCommonDataModel import (
@@ -20,49 +18,15 @@ from vtkmodules.vtkCommonDataModel import (
 )
 from vtkmodules.vtkFiltersCore import vtkAppendPolyData, vtkExtractCells, vtkPolyDataNormals, vtkPolyDataTangents
 from vtkmodules.vtkFiltersTexture import vtkTextureMapToPlane
-from vtkmodules.vtkIOImage import vtkJPEGReader, vtkPNGReader
 from vtkmodules.vtkRenderingCore import (
     vtkActor,
     vtkDataSetMapper,
     vtkPropAssembly,
-    vtkTexture,
 )
 
 from vibra import TEXTURE_DIR, app
 from vibra.engine.mesher.mesh import Mesh
-from vibra.utils.polydata_utils import fill_array
-
-
-def create_vtk_id_list(id_list: Sequence[int]) -> vtkIdList:
-    vtk_id_list = vtkIdList()
-    for id in id_list:
-        vtk_id_list.InsertNextId(id)
-    return vtk_id_list
-
-
-def read_texture(path: str | Path | None):
-    path = Path(path)
-
-    if not path.exists():
-        raise FileNotFoundError(f'Texture file "{path}" not found')
-
-    if path.suffix == ".png":
-        reader = vtkPNGReader()
-    elif path.suffix == ".jpg":
-        reader = vtkJPEGReader()
-    else:
-        raise ValueError(f"Unsupported image format {path.suffix}")
-
-    reader.SetFileName(path)
-    reader.Update()
-
-    texture = vtkTexture()
-    texture.InterpolateOn()
-    texture.RepeatOn()
-    texture.SetInputData(reader.GetOutput())
-    texture.Update()
-
-    return texture
+from vibra.utils.vtk_utils import fill_array, read_texture
 
 
 def get_first_visible_volume(volumes):
